requirements.txt
ExifRead==2.1.2
Flask==2.1.2
mypy==1.1.1
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.3.0
pyproj==3.4.0
scikit-learn==1.1.3
scipy==1.9.3
sympy==1.10.1
utm==0.7.0
Werkzeug==2.2.2